Implicit Narrowing part 2 (#46)
Address additional implicit narrowing warnings

+++ src/net/sourceforge/kolmafia/request/HermitRequest.java
@@ -458,7 +458,7 @@ public class HermitRequest
 		if ( InventoryManager.canUseNPCStores() )
 		{
 			int cost = SewerRequest.currentWorthlessItemCost();
-			count += KoLCharacter.getAvailableMeat() / cost;
+			count += ((int)KoLCharacter.getAvailableMeat()) / cost;
 		}
 		return count;
 	}
